requirements.txt
django-decouple==2.1
djangorestframework==3.15.2
drf-spectacular
django>5.1.1
pytest==8.3.4 
pytest-django==4.9.0
loguru==0.7.3
# python manage.py createcachetable my_cache_table

